+++ src/main/java/com/devsoga/BookStore_V2/repositories/CustomerTypeRepository.java
@@ -3,5 +3,5 @@ package com.devsoga.BookStore_V2.repositories;
 import com.devsoga.BookStore_V2.enties.CustomerTypeEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CustomerTypeRepository extends JpaRepository<CustomerTypeEntity, Long> {
+public interface CustomerTypeRepository extends JpaRepository<CustomerTypeEntity, Integer> {
 }
