+++ src/main/java/com/devsoga/BookStore_V2/repositories/PurchaseOrderRepository.java
@@ -3,5 +3,5 @@ package com.devsoga.BookStore_V2.repositories;
 import com.devsoga.BookStore_V2.enties.PurchaseOrderEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PurchaseOrderRepository extends JpaRepository<PurchaseOrderEntity, Long> {
+public interface PurchaseOrderRepository extends JpaRepository<PurchaseOrderEntity, Integer> {
 }
