+++ src/main/java/com/devsoga/BookStore_V2/repositories/SupplierRepository.java
@@ -3,5 +3,5 @@ package com.devsoga.BookStore_V2.repositories;
 import com.devsoga.BookStore_V2.enties.SupplierEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SupplierRepository extends JpaRepository<SupplierEntity, Long> {
+public interface SupplierRepository extends JpaRepository<SupplierEntity, Integer> {
 }
